Mistake in errorCounter fixed. ~5 FPS
The errorCounter variable was uint8_t type so the maximum value which it can store is 255. There are more than 255 errors in the transference so the variable value was reseted to zero.
The errorCounter is uint16_t type now. The transference times have been modified to get zero errors, and the FPS finally is 5 FPS roughly.