Optical flow frame used fixed.
While I am writing the documentation of Srinivasan algorithm, I have noticed the frame used to get the reference frames is the first one and in the code I was using the second one, so I fix it.
Also the indexes to f2 and f1 has been permuted to be agree with Srinivasan article.